+++ src/components/Modal/index.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+
+const modalRoot = document.getElementById('modal-root'); 
+export class Modal extends React.Component { 
+
+  render() {
+    return ReactDOM.createPortal(
+      this.props.children,
+      modalRoot
+    );
+  }
+
+}
