+++ src/index.html
@@ -1,14 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Connect</title>
 </head>
-
 <body>
   <div id="app"></div>
+  <div id="modal-root"></div>
 </body>
 
 </html>
