+++ src/interfaces/interfaces.ts
@@ -1,0 +1,22 @@
+export interface propsGeneral {
+  data?: {
+    loading?: boolean,
+    getAllProjects?:object[]
+  }
+}
+
+export interface employee { 
+  _id?: string
+  firstName?: string,
+  lastName?: string,
+  position?: string,
+  salary?:number,
+  state?: boolean,
+  project?: Project
+}
+
+export interface Project {
+  _id ?: string;
+  name?: string;
+  department?: string;
+}
