--- src/interfaces/projects.ts
@@ -1,5 +1,0 @@
-export interface Project {
-  _id ?: string;
-  name?: string;
-  department?: string;
-}
